@2.0.0 @Implement internationalization hook and update components to use it
