+++ src/presentation/components/Logo.tsx
@@ -1,16 +1,22 @@
+'use client';
+
 import IconCodingFlavour from "@public/icons/icon-coding-flavour.svg";
 import Component from "@src/data/Models/Component";
+import useI18N from "@src/hooks/useI18N";
 import styles from "@src/presentation/styles/components/_logo.module.scss";
 import Image from "next/image";
 import Link from "next/link";
+import React from "react";
 
 const { logo, logo__image, logo__company } = styles;
 
-const Logo: Component = async ({
+const Logo: Component = ({
   dict
 }) => {
+  const { languageActive } = useI18N();
+
   return (
-    <Link href="/" className={logo} data-testid={"logo"}>
+    <Link href={`/${languageActive}`} className={logo} data-testid={"logo"}>
       <Image
         className={logo__image}
         src={IconCodingFlavour}
