+++ src/presentation/components/Visit.tsx
@@ -1,4 +1,7 @@
+'use client';
+
 import IconArrowUpRight from "@public/icons/icon-arrow-up-right.svg";
+import useI18N from "@src/hooks/useI18N";
 import styles from "@src/presentation/styles/components/_visit.module.scss";
 import Image from "next/image";
 import Link from "next/link";
@@ -13,20 +16,24 @@ export interface IVisitProps {
   href: string;
   target?: "_self" | "_blank" | "_parent" | "_top";
   size?: Size;
+  isExternal?: boolean;
 }
-
 
 const Visit: React.FC<IVisitProps> = ({
   text,
   href,
   target = "_blank",
   size = "regular",
+  isExternal
 }) => {
+  const { languageActive } = useI18N();
+
   const linkClassName = `${visit} ${size === "small" ? smallVisit : ""}`;
+  const finalHref = isExternal ? href : `/${languageActive}${href}`;
 
   return (
     <div className={linkClassName} data-testid={"visit"}>
-      <Link className={visit__link} href={href} target={target} data-testid={`visit-link`}>
+      <Link className={visit__link} href={finalHref} target={target} data-testid={`visit-link`}>
         <span data-testid={"visit-text"} className={visit__text}>
           {text}
         </span>
@@ -43,4 +50,4 @@ const Visit: React.FC<IVisitProps> = ({
   );
 };
 
-export default React.memo(Visit);
+export default Visit
